Added heartbeat
Client(s) now send heart beat(s).
If server doesnt receive a heartbeat in 10 seconds the client is set to "offline".
Offline clients can not be transfered to, they will be switched off of if they were the active pc. Offline pc will interrupr the chain of transfer and block transfers. If the client comes back online and sends hearbeat again they will be set to online.